requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
fastapi==0.79.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
mangum==0.15.0
numpy==1.23.1
pandas==1.4.3
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.2.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
